Current and position sensor working

+++ Core/Inc/FOC.h
@@ -11,7 +11,7 @@
 #include "stm32h7xx_hal.h"
 
 
-void rawdata_to_angle(int32_t rawdata,int32_t &mech_angle,int32_t &electrical_angle,int32_t debug_offset,int32_t pole_pairs);
+void rawdata_to_angle(int32_t rawdata,float &mech_angle,int32_t &electrical_angle,int32_t debug_offset,int32_t pole_pairs);
 void dqz(float Ia,float Ib,float Ic,int32_t theta,float &Iq,float &Id );
 void inv_dqz(float &Va,float &Vb,float &Vc,int32_t theta,float Vq,float Vd );
 #endif /* FOC_H_ */
